Auth/requirements.txt
cassandra-driver

pydantic[email]
fastapi
uvicorn[standard]

python-jose[cryptography]
passlib[bcrypt]
python-multipart
